Xing magic.
Properly compute bitrate & duration for files with Xing headers that only include the frame count and not the file size. In the case I had, the first frame was 32 kbit/s, but all the subsequent frames were 192 kbit/s, and we were using that first bitrate to calculate the duration.